refactor(jei): 替换容器升级类别，移除旧升级类别
- 删除 ShulkerContainerUpgrade 和 HyperdimensionStorageStationUpgrade 相关类别及配方类
- 新增 ContainerUpgradeCategory 类，支持箱子升级与存储站升级两种类型
- 修改 AnvilCraftJeiPlugin，使用 ContainerUpgrade 取代旧的升级类别和配方
- 更新本地化语言文件，替换对应类别的文本条目
- 精简 JEI 注册流程，整合升级类别为统一处理
