+++ src/main/java/dev/dubhe/anvilcraft/data/lang/JeiLang.java
@@ -83,15 +83,11 @@ public class JeiLang {
         provider.add("gui.anvilcraft.category.beacon_conversion.activate", "Use this item to activate beacon");
         provider.add("gui.anvilcraft.category.beacon_conversion.beacon_base", "Use this block as beacon base");
 
-        provider.add("gui.anvilcraft.category.shulker_container_upgrade", "Shulker Container Upgrade");
-        provider.add("gui.anvilcraft.category.shulker_container_upgrade.drop_on_top", "Drop these items onto the large crate");
-        provider.add("gui.anvilcraft.category.shulker_container_upgrade.strike", "Strike with an anvil to upgrade the crate");
-
-        provider.add("gui.anvilcraft.category.hyperdimension_storage_station_upgrade", "Hyperdimension Storage Station Upgrade");
-        provider.add("gui.anvilcraft.category.hyperdimension_storage_station_upgrade.drop_on_top", "Drop these items onto the shulker container");
-        provider.add("gui.anvilcraft.category.hyperdimension_storage_station_upgrade.strike", "Strike with an anvil to upgrade the container");
+        provider.add("gui.anvilcraft.category.container_upgrade", "Container Upgrade");
+        provider.add("gui.anvilcraft.category.container_upgrade.drop_on_top", "Drop these items onto the container");
+        provider.add("gui.anvilcraft.category.container_upgrade.strike", "Strike with an anvil to upgrade the container");
         provider.add(
-            "gui.anvilcraft.category.hyperdimension_storage_station_upgrade.requires_expansion",
+            "gui.anvilcraft.category.container_upgrade.requires_expansion",
             "Requires 4 Space Over-compressor Expansions"
         );
