--- src/main/java/dev/dubhe/anvilcraft/integration/jei/recipe/ShulkerContainerUpgradeRecipe.java
@@ -1,9 +1,0 @@
-package dev.dubhe.anvilcraft.integration.jei.recipe;
-
-import com.google.common.collect.ImmutableList;
-
-public record ShulkerContainerUpgradeRecipe() {
-    public static ImmutableList<ShulkerContainerUpgradeRecipe> getAllRecipes() {
-        return ImmutableList.of(new ShulkerContainerUpgradeRecipe());
-    }
-}
